Perl Index Program
A Perl program reading from "indexDOC" document, creating a hash from the document (keys are assigned to one line, the value is assigned to the following line).
Accepts arguments "r" to sort the terms alphabetically and display the entire doc. Argument "u" allows the user to add a new term and description to the doc. Argument "s" allows the user to search the doc for specific terms.